Add a listening socket for spectators
So it's easily possible to follow the game remotely.

Use `-K` to give a key (string) each spectator needs
to send in order to be authenticated. If you don't
give a key, any spectator is allowed - this may give
players the opportunity to do a side attack and get
the whole map.