Allow local address reuse
Do not block the game port when a player hasn't closed it's
side of the socket and the server has just been terminated
and is about to be restarted.

Both sides of a socket need to close() or the port will be
blocked and in TIME_WAIT state for a minute or so. This can
be a minor inconvenience if a player's bot crashed and the
same player terminates the server and tries to start it
again.

For details about this issue see:
https://hea-www.harvard.edu/~fine/Tech/addrinuse.html