Use a ring buffer to track character for word
Simply using an incrementing pointer that walks the searched
word with every command character it gets is insufficient,
unfortunately.

If the searched word is "oof", and a bot sends "ooof", then
the third 'o' won't match the 'f' and the pointer is reset
to the beginning of the searched string. The last 'f' that
the bot did send will be matched against the first 'o' of
the searched string. What fails, of course.

So the walking pointer cannot ignore any garbage in front
of the word that accidently matches the beginning of the
searched string.

But there is more. If the searched word is "abc", and a bot
sends "ababc", the second 'a' is not recognized as a new
beginning because it fails to match the 'c' at that position.

So it's inevitable to save past commands in a ring buffer
of the size of the searched word.