Make sure asteroids are always visible
Because asteroids move independently now, they cannot
be erased and drawn in one loop because the old position
of one asteroid may be the new position of another.

Also, asteroids do not destroy a given map anymore.
Instead, the tile is restored to what it was before
the asteroid occupied it.

And to give players a fair chance of survival, no
asteroid is placed directly beside a player at the
beginning of the game. This is especially important
because asteroids always move first. If an asteroid
is placed directly besides a player, it may hit the
player before the player got any chance to avoid it.