Try to find a free spot to place players
Don't simply overwrite the map when placing players
but try to find a free spot instead. This is especially
important for the maze map type because changing the
map may destroy the maze or place the player in the
middle of a cross of wall where any movement would be
impossible.

And try to find a free spot when placing the exit
field in escape too. Same reason: changing the map
would destroy the maze.