Assign player names by using a map file
The map file can have multiple entries for the same name
making it possible to map a number of IP addresses to a
name. This way, a team may use different machines while
keeping the same name.

Also, if a player or team drops out of an event, the
remaining players can keep their names.