Clarify that a bot must not send multiple commands
Any additional command that is read from the network buffer
within a turn is discarded. Because a turn ends when all
players have moved, bots in games with just one player will
never see that.

The more players there are, the more likely it is that
additional commands are silently discarded instead of
getting processed in subsequent turns.

So to be sure, a bot must never send multiple command
characters in response to a map. You get a map, you send
a byte.