Clean up Python cheatsheet
Resolve mix of tabs and spaces for indentation. You should really
know better.

Remove tabs in the middle of a line. Never do that. A tab is meant
to specify exactly one level of indent from the edge of the document.

Remove trailing white-space. You should really use a better editor.

Use proper headline structure instead of styling something bold.

Use Bots instead of BOTS because that's how it's called in the README.