Set --max-lag to unlimited by default
Because new people will start with the manual bots and this means
some will most probably miss a few turns. It's confusing when
they're disconnected without knowing why.

This flag is supposed to be used to make the server automatically
disconnect defect bots that fail to send a command for a couple of
turns. Not sending a command means that a turn takes a full second
what makes the game slow.